Cosmetic fix

+++ src/string.cpp
@@ -501,8 +501,8 @@ String String::windowsDecode(void) const
 		char chr = at(i);
 		if(chr & 0x80)
 		{
-			result+= char((uint8_t(chr) >> 6) | uint8_t(0xC0));	// 1st byte
-			result+= char((uint8_t(chr) & ~0xC0) | uint8_t(0x80));	// 2nd byte
+			result+= char((uint8_t(chr) >> 6) | uint8_t(0xC0));		// 1st byte
+			result+= char((uint8_t(chr) & ~uint8_t(0xC0) | uint8_t(0x80));	// 2nd byte
 		}
 		else {
 			result+= chr;
